requirements.txt
# Core dependencies
streamlit>=1.30.0
openai>=1.12.0
supabase>=2.0.0
python-dotenv>=1.0.0

# File handling
PyPDF2>=3.0.0
python-docx>=1.0.0
python-magic>=0.4.27

# Utilities
tqdm>=4.66.0
